requirements.txt
peewee==3.9.3
Pillow==6.0.0
beautifulsoup4==4.7.1
jinja2==2.10
matplotlib